修改ISgmt中的z1、z2数据类型double to float，删去结构体Sline2中的p，改facetId类型UINT64 to UINT

+++ CAD/CADDoc.cpp
@@ -657,7 +657,7 @@ double CCADDoc::GetMeanZ()
 }
 
 // nt add 2017/5/27
-void CCADDoc::CutAll(double z)
+void CCADDoc::CutAll(float z) // 2022/08/16 smf modify: double to float
 {
 	for( STL* stl = m_stls ; stl ; stl = stl->next )
 	{
