+++ CAD/MainFrm.cpp
@@ -351,7 +351,7 @@ void CMainFrame::OnSlider1()
 		cadPromptStr(_T("无STL模型!")) ;
 		return ;
 	}
-	double z = pDoc->m_iLayer*pDoc->GetZMax()/pDoc->m_nLayer ;
+	float z = (float)(pDoc->m_iLayer*pDoc->GetZMax()/pDoc->m_nLayer) ; // 2022/08/16 smf modify: 强转double to float
 	pDoc->CutAll(z) ;
 	pView->m_drawSection = 1 ;
 	pView->m_z = z ;
